+++ bin/faucet/src/server/mod.rs
@@ -1,7 +1,7 @@
 use std::{
     collections::HashSet,
     convert::Infallible,
-    time::{SystemTime, UNIX_EPOCH},
+    time::{Duration, SystemTime, UNIX_EPOCH},
 };
 
 use anyhow::Context;
@@ -58,6 +58,7 @@ impl Server {
         asset_options: AssetOptions,
         request_sender: RequestSender,
         pow_secret: &str,
+        pow_challenge_lifetime: Duration,
         api_keys: &[ApiKey],
     ) -> Self {
         let mint_state = GetTokensState::new(request_sender, asset_options.clone());
@@ -73,7 +74,7 @@ impl Server {
         hasher.update(pow_secret.as_bytes());
         let secret_bytes: [u8; 32] = hasher.finalize().into();
 
-        let pow = PoW::new(secret_bytes);
+        let pow = PoW::new(secret_bytes, pow_challenge_lifetime);
 
         Server {
             mint_state,